+++ CHANGELOG.md
@@ -6,12 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.2] - 2023-08-01
+### Changed
+- `Relation::filter_columns`: add a conversion of the arguments
+
 ## [0.2.1] - 2023-07-26
 ### Added
 - join utils [MR72](https://github.com/Qrlew/qrlew/pull/72)
 - `Relation::possible_values` and used it in thresholding [MR73](https://github.com/Qrlew/qrlew/pull/73)
 ### Changed
-- `Relation::filter_columns`: changed siganture and add a conversion
+- `Relation::filter_columns`: changed signature
 -  join utils [MR72](https://github.com/Qrlew/qrlew/pull/72)
 -  Fixed table naming